修复5.0以下系统资源分区方案插件找不到AppCompat主题的问题
版本:9.9.0
描述:原因是5.0以下，系统方法创建的AssetManager不支持资源表扩充，需要new

+++ SdkLibrary/src/org/qiyi/pluginlibrary/utils/MultiDexExtractor.java
@@ -293,7 +293,7 @@ class MultiDexExtractor {
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong(keyPrefix + KEY_TIME_STAMP, timeStamp);
         editor.putLong(keyPrefix + KEY_CRC, crc);
-        editor.putLong(keyPrefix + KEY_DEX_NUMBER, dexFiles.size() + 1L);
+        editor.putInt(keyPrefix + KEY_DEX_NUMBER, dexFiles.size() + 1);
 
         int dexNumber = 2;
         for (ExtractedDex dex : dexFiles) {
